feat: implement migration up command
