+++ internal/migration/apply/apply.go
@@ -22,24 +22,28 @@ func MigrateDatabase(ctx context.Context, conn *pgx.Conn, fsys afero.Fs) error {
 	if err != nil {
 		return err
 	}
-	// Apply migrations
-	for _, filename := range migrations {
-		if err := migrateUp(ctx, conn, filename, fsys); err != nil {
+	return MigrateUp(ctx, conn, migrations, fsys)
+}
+
+func MigrateUp(ctx context.Context, conn *pgx.Conn, pending []string, fsys afero.Fs) error {
+	if err := repair.CreateMigrationTable(ctx, conn); err != nil {
+		return err
+	}
+	for _, filename := range pending {
+		if err := applyMigration(ctx, conn, filename, fsys); err != nil {
 			return err
 		}
 	}
 	return nil
 }
 
-func migrateUp(ctx context.Context, conn *pgx.Conn, filename string, fsys afero.Fs) error {
+func applyMigration(ctx context.Context, conn *pgx.Conn, filename string, fsys afero.Fs) error {
 	fmt.Fprintln(os.Stderr, "Applying migration "+utils.Bold(filename)+"...")
 	path := filepath.Join(utils.MigrationsDir, filename)
 	migration, err := NewMigrationFromFile(path, fsys)
 	if err != nil {
 		return err
 	}
-	// Skip inserting to migration history table
-	migration.version = ""
 	return migration.ExecBatch(ctx, conn)
 }
